chore: rename backend_source_build.rs
The module reflects the type name now
